+++ build_runner/test/generate/build_configuration_test.dart
@@ -3,24 +3,21 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:build/build.dart';
-import 'package:build_runner/src/package_graph/apply_builders.dart';
+import 'package:build_test/build_test.dart';
 import 'package:test/test.dart';
-
-import '../common/common.dart';
 
 void main() {
   test('uses builder options', () async {
     Builder copyBuilder(BuilderOptions options) => TestBuilder(
       buildExtensions: replaceExtension(
-        options.config['inputExtension'] as String,
+        options.config['inputExtension'] as String? ?? '',
         '.copy',
       ),
+      name: 'a:optioned_builder',
     );
 
-    await testPhases(
-      [
-        apply('a:optioned_builder', [copyBuilder], toRoot(), hideOutput: false),
-      ],
+    await testBuilderFactories(
+      [copyBuilder],
       {
         'a|lib/file.nomatch': 'a',
         'a|lib/file.matches': 'b',
@@ -33,10 +30,10 @@ targets:
           inputExtension: .matches
 ''',
       },
+      testingBuilderConfig: false,
       outputs: {'a|lib/file.copy': 'b'},
     );
   });
-
   test('isRoot is applied correctly', () async {
     Builder copyBuilder(BuilderOptions options) => TestBuilder(
       buildExtensions: replaceExtension(
@@ -44,22 +41,10 @@ targets:
         options.isRoot ? '.root.copy' : '.dep.copy',
       ),
     );
-    var packageGraph = buildPackageGraph({
-      rootPackage('a'): ['b'],
-      package('b'): [],
-    });
-    await testPhases(
-      [
-        apply(
-          'a:optioned_builder',
-          [copyBuilder],
-          toAllPackages(),
-          hideOutput: true,
-        ),
-      ],
+    await testBuilderFactories(
+      [copyBuilder],
       {'a|lib/a.txt': 'a', 'b|lib/b.txt': 'b'},
-      outputs: {r'$$a|lib/a.root.copy': 'a', r'$$b|lib/b.dep.copy': 'b'},
-      packageGraph: packageGraph,
+      outputs: {r'a|lib/a.root.copy': 'a', r'b|lib/b.dep.copy': 'b'},
     );
   });
 }